Refactored code to move dcr to oidc module
Instead of having types declared in the core module, and code in the browser module, all the oidc-specific code regarding dynamic client registration moved to the oidc module. At this point, the tests have not been updated.
