+++ packages/oidc-dpop-client-browser/package.json
@@ -31,7 +31,9 @@
     "rollup-plugin-typescript2": "^0.27.1"
   },
   "dependencies": {
-    "oidc-client": "^1.10.1"
+    "@types/url-parse": "^1.4.3",
+    "oidc-client": "^1.10.1",
+    "url-parse": "^1.4.7"
   },
   "publishConfig": {
     "access": "public"
