+++ packages/oidc-dpop-client-browser/src/IIssuerConfig.ts
