+++ packages/oidc-dpop-client-browser/tsconfig.json
@@ -10,5 +10,6 @@
     "out": "website/docs/api/oidc-dpop-client-browser"
   },
 
-  "include": ["src/**/*"]
+  "include": ["src/**/*"],
+  "exclude": ["*.spec.ts"]
 }
